chore: check in angular docs
