+++ packages/react/src/components/diagrams/0_Diagrams.stories.js
@@ -55,7 +55,8 @@ stories.add('Start here', () => {
 
 			<p {...paragraphStyle}>
 				Note that Carbon Charts does not provide layouts for diagrams.
-				You can utilize these components alongside graphing libraries, or by composing your own layouts.
+				You can utilize these components alongside graphing libraries,
+				or by composing your own layouts.
 			</p>
 
 			<h2 {...h2Style}>Examples</h2>
